Fix emitting GPG_ERR_MINUS_ONE on 32 bit architectures.
  - Fixes 54ab16247ff1be333bdf757a324ee02a660e4742.

  - Fixes #44.

+++ src/status.rs
@@ -820,7 +820,7 @@ impl Status<'_> {
                 location,
                 error,
             } => {
-                writeln!(w, "FAILURE {} {}", location, *error as isize)?;
+                writeln!(w, "FAILURE {} {}", location, *error as u32)?;
             },
 
             Unexpected(reason) => {
